requirements.txt
# Core ML and Computer Vision
tensorflow==2.15.0
opencv-python==4.8.1.78
numpy==1.26.4
pillow==10.2.0

# Streamlit and Visualization
streamlit==1.31.1
plotly==5.18.0
pandas==2.2.0
matplotlib==3.8.2

# Authentication and Security
streamlit-authenticator==0.3.1
bcrypt==4.1.2
pyjwt==2.8.0

# Testing
pytest==8.0.0
hypothesis==6.98.3
pytest-cov==4.1.0

# Model Optimization
tensorflow-model-optimization==0.8.0

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1
scikit-learn==1.4.0
